+++ db/finance.go
@@ -26,8 +26,6 @@ type BankTransaction struct {
 
 // FinTransMonthly reads all financial transactions from given month.
 func (c *Client) FinTransMonthly(year int, month int) ([]BankTransaction, error) {
-	startTs := time.Now()
-	transactions := make([]BankTransaction, 0, 500)
 	log.Info().Msgf("[%s] start reading transactions for [%d-%d-*]", dbFinPrefix, year, month)
 
 	monthStr := strconv.Itoa(month)
@@ -35,43 +33,8 @@ func (c *Client) FinTransMonthly(year int, month int) ([]BankTransaction, error)
 		monthStr = "0" + monthStr
 	}
 	likeCond := fmt.Sprintf("%d-%s-%%", year, monthStr)
-	log.Info().Msgf("[%s] transactions will be filtered with [LIKE '%s']", dbFinPrefix, likeCond)
 
-	rows, qErr := c.dbConn.Query(monthlyTransactionQuery(), likeCond)
-	if qErr != nil {
-		log.Error().Err(qErr).Msgf("[%s] monthlyTransactionQuery failed", dbFinPrefix)
-		return transactions, qErr
-	}
-
-	var id int
-	var amount, endingBalanceValue float64
-	var accNumber, execDate, orderDate, amountCurr, description string
-	var ttype, endingBalanceCurr *string
-	for rows.Next() {
-		sErr := rows.Scan(&id, &accNumber, &execDate, &orderDate, &ttype, &amountCurr, &amount,
-			&endingBalanceCurr, &endingBalanceValue, &description)
-		if sErr != nil {
-			log.Warn().Err(sErr).Msgf("[%s] while scanning results of monthlyTransactionQuery", dbFinPrefix)
-			continue
-		}
-		endingBalanceCopy := endingBalanceValue
-		transactions = append(transactions, BankTransaction{
-			TransactionId:         id,
-			AccountNumber:         accNumber,
-			ExecutionDate:         execDate,
-			OrderDate:             orderDate,
-			Type:                  ttype,
-			AmountCurrency:        amountCurr,
-			AmountValue:           amount,
-			EndingBalanceCurrency: endingBalanceCurr,
-			EndingBalanceValue:    &endingBalanceCopy,
-			Description:           description,
-		})
-	}
-	log.Info().Int("rowsLoaded", len(transactions)).Dur("duration", time.Since(startTs)).
-		Msgf("[%s] finished reading transactions for [%d-%d-*]", dbDocsPrefix, year, month)
-
-	return transactions, nil
+	return c.finQueryTransactions(monthlyTransactionQuery(), likeCond)
 }
 
 // FinInsertTransactions inserts bank transactions into database. Transactions
@@ -116,6 +79,55 @@ func (c *Client) FinInsertTransactions(transactions []BankTransaction) error {
 	return nil
 }
 
+// FinTransByDescription reads all financial transaction that contains given
+// phrase in the transaction description.
+func (c *Client) FinTransByDescription(phrase string) ([]BankTransaction, error) {
+	likeCond := fmt.Sprintf("%%%s%%", phrase)
+	return c.finQueryTransactions(finTransactionByDescriptionQuery(), likeCond)
+}
+
+func (c *Client) finQueryTransactions(query string, args ...interface{}) ([]BankTransaction, error) {
+	startTs := time.Now()
+	transactions := make([]BankTransaction, 0, 500)
+	log.Info().Msgf("[%s] start reading financial transactions", dbFinPrefix)
+
+	rows, qErr := c.dbConn.Query(query, args...)
+	if qErr != nil {
+		log.Error().Err(qErr).Msgf("[%s] financial query failed", dbFinPrefix)
+		return transactions, qErr
+	}
+
+	var id int
+	var amount, endingBalanceValue float64
+	var accNumber, execDate, orderDate, amountCurr, description string
+	var ttype, endingBalanceCurr *string
+	for rows.Next() {
+		sErr := rows.Scan(&id, &accNumber, &execDate, &orderDate, &ttype, &amountCurr, &amount,
+			&endingBalanceCurr, &endingBalanceValue, &description)
+		if sErr != nil {
+			log.Warn().Err(sErr).Msgf("[%s] while scanning results of [%s] query", dbFinPrefix, query)
+			continue
+		}
+		endingBalanceCopy := endingBalanceValue
+		transactions = append(transactions, BankTransaction{
+			TransactionId:         id,
+			AccountNumber:         accNumber,
+			ExecutionDate:         execDate,
+			OrderDate:             orderDate,
+			Type:                  ttype,
+			AmountCurrency:        amountCurr,
+			AmountValue:           amount,
+			EndingBalanceCurrency: endingBalanceCurr,
+			EndingBalanceValue:    &endingBalanceCopy,
+			Description:           description,
+		})
+	}
+	log.Info().Int("rowsLoaded", len(transactions)).Dur("duration", time.Since(startTs)).
+		Msgf("[%s] finished reading financial transactions", dbFinPrefix)
+
+	return transactions, nil
+}
+
 // Inserts single bank transaction into database.
 func (c *Client) insertTransaction(t BankTransaction, tx *sql.Tx) error {
 	_, insErr := tx.Exec(
@@ -157,5 +169,25 @@ func insertTransactionQuery() string {
 		AmountValue, EndingBalanceCurrency, EndingBalanceValue, Description
 	)
 	VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
+	`
+}
+
+func finTransactionByDescriptionQuery() string {
+	return `
+	SELECT
+		TransactionId,
+		AccountNumber,
+		ExecutionDate,
+		OrderDate,
+		TType,
+		AmountCurrency,
+		AmountValue,
+		EndingBalanceCurrency,
+		EndingBalanceValue,
+		Description
+	FROM
+		bankTransactions
+	WHERE
+		Description LIKE ?
 	`
 }